ExynosOverlayDisplay: Pass the same parameter type to std::min
* std::min is a function template on T, which is the type of both params.
* 'NUM_HW_WINDOWS - 1' is a type of size_t, so cast it to uint32_t to comply

+++ libdisplay/ExynosOverlayDisplay.cpp
@@ -968,7 +968,7 @@ void ExynosOverlayDisplay::determineBandwidthSupport(hwc_display_contents_1_t *c
             pixel_used[dma_ch_idx] += pixels_needed;
             win_idx++;
             win_idx = (win_idx == mFirstFb) ? (win_idx + 1) : win_idx;
-            win_idx = min(win_idx, NUM_HW_WINDOWS - 1);
+            win_idx = min(win_idx, static_cast<uint32_t>(NUM_HW_WINDOWS - 1));
             windows_left--;
             if (gsc_required) {
                 mGscUsed = true;
